feat(auth): implement signup flow with validation, mapping, and security foundation
add User entity with fields, role enum, and timestamps

add SignUpRequest and SignupResponse DTOs with validation

add MapStruct UserMapper for request to entity mapping

implement UserRepository with existsByEmail and findByEmail

implement AuthService and signup logic with password encoding and default role

add ApiResponse wrapper for standardized responses

add AuthController with /auth/register endpoint

add logging and transactional boundary for signup

+++ src/main/resources/db/migration/V1__create_categories_table.sql
@@ -6,7 +6,7 @@ CREATE TABLE categories
     -- GENERATED ALWAYS AS IDENTITY
     -- PostgreSQL сам генерує унікальний ID
     -- Це краще ніж вручну задавати значення
-    category_id INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    category_id BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
 
     -- Назва категорії
     -- NOT NULL, бо категорія без назви не має сенсу
