+++ src/main/resources/db/migration/V5__create_cart_items_table.sql
@@ -4,7 +4,7 @@
 CREATE TABLE orders
 (
     -- Унікальний ідентифікатор
-    order_id INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    order_id BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
 
     -- Користувач, який зробив замовлення
     user_id INT NOT NULL,
